fix(pnpm): defer lock file constraint parsing
Closes #22250
