+++ lib/modules/manager/npm/extract/pnpm.ts
@@ -1,7 +1,6 @@
 import is from '@sindresorhus/is';
 import { findPackages } from 'find-packages';
 import { load } from 'js-yaml';
-import semver from 'semver';
 import upath from 'upath';
 import { GlobalConfig } from '../../../../config/global';
 import { logger } from '../../../../logger';
@@ -191,101 +190,3 @@ export async function getPnpmLock(filePath: string): Promise<LockFile> {
     return { lockedVersions: {} };
   }
 }
-
-export function getConstraints(
-  lockfileVersion: number,
-  constraints?: string
-): string {
-  let newConstraints = constraints;
-
-  // find matching lockfileVersion and use its constraints
-  // if no match found use lockfileVersion 5
-  // lockfileVersion 5 is the minimum version required to generate the pnpm-lock.yaml file
-  const { lowerBound, upperBound, lowerConstraint, upperConstraint } =
-    lockToPnpmVersionMapping.find(
-      (m) => m.lockfileVersion === lockfileVersion
-    ) ?? {
-      lockfileVersion: 5.0,
-      lowerBound: '2.24.0',
-      upperBound: '3.5.0',
-      lowerConstraint: '>=3',
-      upperConstraint: '<3.5.0',
-    };
-
-  // inorder to ensure that the constraint doesn't allow any pnpm versions that can't generate the extracted lockfileVersion
-  // compare the current constraint to the lowerBound and upperBound of the lockfileVersion
-  // if the current constraint is not comaptible, add the lowerConstraint and upperConstraint, whichever is needed
-  if (newConstraints) {
-    // if constraint satisfies versions lower than lowerBound add the lowerConstraint to narrow the range
-    if (semver.satisfies(lowerBound, newConstraints)) {
-      newConstraints += ` ${lowerConstraint}`;
-    }
-
-    // if constraint satisfies versions higher than upperBound add the upperConstraint to narrow the range
-    if (
-      upperBound &&
-      upperConstraint &&
-      semver.satisfies(upperBound, newConstraints)
-    ) {
-      newConstraints += ` ${upperConstraint}`;
-    }
-  }
-  // if no constraint is present, add the lowerConstraint and upperConstraint corresponding to the lockfileVersion
-  else {
-    newConstraints = `${lowerConstraint}${
-      upperConstraint ? ` ${upperConstraint}` : ''
-    }`;
-  }
-
-  return newConstraints;
-}
-
-/**
- pnpm lockfiles have corresponding version numbers called "lockfileVersion"
- each lockfileVersion can only be generated by a certain pnpm version ranges
- eg. lockfileVersion: 5.4 can only be generated by pnpm version >=7 && <8
- official list can be found here : https:github.com/pnpm/spec/tree/master/lockfile
- we use the mapping present below to find the compatible pnpm version range for a given lockfileVersion
-
- the various terms used in the mapping are explained below:
- lowerConstriant : lowest pnpm version that can generate the lockfileVersion
- upperConstraint : highest pnpm version that can generate the lockfileVersion
- lowerBound      : highest pnpm version that is less than the lowerConstraint
- upperBound      : lowest pnpm version that is greater than upperConstraint
-
- For handling future lockfileVersions, we need to:
- 1. add a upperBound and upperConstraint to the current lastest lockfileVersion
- 2. add an object for the new lockfileVersion with lowerBound and lowerConstraint
- */
-
-const lockToPnpmVersionMapping = [
-  { lockfileVersion: 6.0, lowerBound: '7.32.0', lowerConstraint: '>=8' },
-  {
-    lockfileVersion: 5.4,
-    lowerBound: '6.35.1',
-    upperBound: '8.0.0',
-    lowerConstraint: '>=7',
-    upperConstraint: '<8',
-  },
-  {
-    lockfileVersion: 5.3,
-    lowerBound: '5.18.10',
-    upperBound: '7.0.0',
-    lowerConstraint: '>=6',
-    upperConstraint: '<7',
-  },
-  {
-    lockfileVersion: 5.2,
-    lowerBound: '5.9.3',
-    upperBound: '5.18.10',
-    lowerConstraint: '>=5.10.0',
-    upperConstraint: '<6',
-  },
-  {
-    lockfileVersion: 5.1,
-    lowerBound: '3.4.1',
-    upperBound: '5.9.3',
-    lowerConstraint: '>=3.5.0',
-    upperConstraint: '<5.9.3',
-  },
-];
