+++ lib/modules/manager/npm/post-update/pnpm.ts
@@ -153,3 +153,63 @@ async function getPnpmConstraint(
   }
   return result;
 }
+
+export function getConstraintsFromLockFile(lockfileVersion: number): string {
+  // find matching lockfileVersion and use its constraints
+  // if no match found use lockfileVersion 5
+  // lockfileVersion 5 is the minimum version required to generate the pnpm-lock.yaml file
+  const { lowerConstraint, upperConstraint } = lockToPnpmVersionMapping.find(
+    (m) => m.lockfileVersion === lockfileVersion
+  ) ?? {
+    lockfileVersion: 5.0,
+    lowerConstraint: '>=3',
+    upperConstraint: '<3.5.0',
+  };
+  const newConstraints = `${lowerConstraint}${
+    upperConstraint ? ` ${upperConstraint}` : ''
+  }`;
+
+  return newConstraints;
+}
+
+/**
+ pnpm lockfiles have corresponding version numbers called "lockfileVersion"
+ each lockfileVersion can only be generated by a certain pnpm version ranges
+ eg. lockfileVersion: 5.4 can only be generated by pnpm version >=7 && <8
+ official list can be found here : https:github.com/pnpm/spec/tree/master/lockfile
+ we use the mapping present below to find the compatible pnpm version range for a given lockfileVersion
+
+ the various terms used in the mapping are explained below:
+ lowerConstriant : lowest pnpm version that can generate the lockfileVersion
+ upperConstraint : highest pnpm version that can generate the lockfileVersion
+ lowerBound      : highest pnpm version that is less than the lowerConstraint
+ upperBound      : lowest pnpm version that is greater than upperConstraint
+
+ For handling future lockfileVersions, we need to:
+ 1. add a upperBound and upperConstraint to the current lastest lockfileVersion
+ 2. add an object for the new lockfileVersion with lowerBound and lowerConstraint
+ */
+
+const lockToPnpmVersionMapping = [
+  { lockfileVersion: 6.0, lowerConstraint: '>=8' },
+  {
+    lockfileVersion: 5.4,
+    lowerConstraint: '>=7',
+    upperConstraint: '<8',
+  },
+  {
+    lockfileVersion: 5.3,
+    lowerConstraint: '>=6',
+    upperConstraint: '<7',
+  },
+  {
+    lockfileVersion: 5.2,
+    lowerConstraint: '>=5.10.0',
+    upperConstraint: '<6',
+  },
+  {
+    lockfileVersion: 5.1,
+    lowerConstraint: '>=3.5.0',
+    upperConstraint: '<5.9.3',
+  },
+];
